Shifting otel related methods to its own file
