+++ pkg/gofr/gofr.go
@@ -16,20 +16,11 @@ import (
 	"time"
 
 	"github.com/gorilla/mux"
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
-	"go.opentelemetry.io/otel/exporters/zipkin"
-	"go.opentelemetry.io/otel/propagation"
-	"go.opentelemetry.io/otel/sdk/resource"
-	sdktrace "go.opentelemetry.io/otel/sdk/trace"
-	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
 	"golang.org/x/sync/errgroup"
 
-	"gofr.dev/pkg/gofr/cmd/terminal"
 	"gofr.dev/pkg/gofr/config"
 	"gofr.dev/pkg/gofr/container"
 	gofrHTTP "gofr.dev/pkg/gofr/http"
-	"gofr.dev/pkg/gofr/http/middleware"
 	"gofr.dev/pkg/gofr/logging"
 	"gofr.dev/pkg/gofr/metrics"
 	"gofr.dev/pkg/gofr/migration"
@@ -55,83 +46,6 @@ type App struct {
 	httpRegistered bool
 
 	subscriptionManager SubscriptionManager
-}
-
-// New creates an HTTP Server Application and returns that App.
-func New() *App {
-	app := &App{}
-	app.readConfig(false)
-	app.container = container.NewContainer(app.Config)
-
-	app.initTracer()
-
-	// Metrics Server
-	port, err := strconv.Atoi(app.Config.Get("METRICS_PORT"))
-	if err != nil || port <= 0 {
-		port = defaultMetricPort
-	}
-
-	if !isPortAvailable(port) {
-		app.container.Logger.Fatalf("metrics port %d is blocked or unreachable", port)
-	}
-
-	app.metricServer = newMetricServer(port)
-
-	// HTTP Server
-	port, err = strconv.Atoi(app.Config.Get("HTTP_PORT"))
-	if err != nil || port <= 0 {
-		port = defaultHTTPPort
-	}
-
-	app.httpServer = newHTTPServer(app.container, port, middleware.GetConfigs(app.Config))
-	app.httpServer.certFile = app.Config.GetOrDefault("CERT_FILE", "")
-	app.httpServer.keyFile = app.Config.GetOrDefault("KEY_FILE", "")
-
-	// Add Default routes
-	app.add(http.MethodGet, "/.well-known/health", healthHandler)
-	app.add(http.MethodGet, "/.well-known/alive", liveHandler)
-	app.add(http.MethodGet, "/favicon.ico", faviconHandler)
-
-	app.checkAndAddOpenAPIDocumentation()
-
-	if app.Config.Get("APP_ENV") == "DEBUG" {
-		app.httpServer.RegisterProfilingRoutes()
-	}
-
-	// gRPC Server
-	port, err = strconv.Atoi(app.Config.Get("GRPC_PORT"))
-	if err != nil || port <= 0 {
-		port = defaultGRPCPort
-	}
-
-	app.grpcServer = newGRPCServer(app.container, port)
-
-	app.subscriptionManager = newSubscriptionManager(app.container)
-
-	// static file server
-	currentWd, _ := os.Getwd()
-	checkDirectory := filepath.Join(currentWd, defaultPublicStaticDir)
-
-	if _, err = os.Stat(checkDirectory); err == nil {
-		app.AddStaticFiles(defaultPublicStaticDir, checkDirectory)
-	}
-
-	return app
-}
-
-// NewCMD creates a command-line application.
-func NewCMD() *App {
-	app := &App{}
-	app.readConfig(true)
-	app.container = container.NewContainer(nil)
-	app.container.Logger = logging.NewFileLogger(app.Config.Get("CMD_LOGS_FILE"))
-	app.cmd = &cmd{
-		out: terminal.New(),
-	}
-	app.container.Create(app.Config)
-	app.initTracer()
-
-	return app
 }
 
 // Run starts the application. If it is an HTTP server, it will start the server.
@@ -402,141 +316,6 @@ func (a *App) Migrate(migrationsMap map[int64]migration.Migrate) {
 	}()
 
 	migration.Run(migrationsMap, a.container)
-}
-
-func (a *App) initTracer() {
-	traceRatio, err := strconv.ParseFloat(a.Config.GetOrDefault("TRACER_RATIO", "1"), 64)
-	if err != nil {
-		a.container.Error(err)
-	}
-
-	tp := sdktrace.NewTracerProvider(
-		sdktrace.WithResource(resource.NewWithAttributes(
-			semconv.SchemaURL,
-			semconv.ServiceNameKey.String(a.container.GetAppName()),
-		)),
-		sdktrace.WithSampler(sdktrace.ParentBased(sdktrace.TraceIDRatioBased(traceRatio))),
-	)
-	otel.SetTracerProvider(tp)
-	otel.SetTextMapPropagator(propagation.NewCompositeTextMapPropagator(propagation.TraceContext{}, propagation.Baggage{}))
-	otel.SetErrorHandler(&otelErrorHandler{logger: a.container.Logger})
-
-	traceExporter := a.Config.Get("TRACE_EXPORTER")
-	tracerURL := a.Config.Get("TRACER_URL")
-
-	// deprecated : tracer_host and tracer_port are deprecated and will be removed in upcoming versions.
-	tracerHost := a.Config.Get("TRACER_HOST")
-	tracerPort := a.Config.GetOrDefault("TRACER_PORT", "9411")
-
-	if !isValidConfig(a.Logger(), traceExporter, tracerURL, tracerHost, tracerPort) {
-		return
-	}
-
-	exporter, err := a.getExporter(traceExporter, tracerHost, tracerPort, tracerURL)
-
-	if err != nil {
-		a.container.Error(err)
-	}
-
-	batcher := sdktrace.NewBatchSpanProcessor(exporter)
-	tp.RegisterSpanProcessor(batcher)
-}
-
-func isValidConfig(logger logging.Logger, name, url, host, port string) bool {
-	if url == "" && name == "" {
-		logger.Debug("tracing is disabled, as configs are not provided")
-		return false
-	}
-
-	if url != "" && name == "" {
-		logger.Error("missing TRACE_EXPORTER config, should be provided with TRACER_URL to enable tracing")
-		return false
-	}
-
-	//nolint:revive // early-return is not possible here, as below is the intentional logging flow
-	if url == "" && name != "" && !strings.EqualFold(name, "gofr") {
-		if host != "" && port != "" {
-			logger.Warn("TRACER_HOST and TRACER_PORT are deprecated, use TRACER_URL instead")
-		} else {
-			logger.Error("missing TRACER_URL config, should be provided with TRACE_EXPORTER to enable tracing")
-			return false
-		}
-	}
-
-	return true
-}
-
-func (a *App) getExporter(name, host, port, url string) (sdktrace.SpanExporter, error) {
-	var (
-		exporter sdktrace.SpanExporter
-		err      error
-	)
-
-	authHeader := a.Config.Get("TRACER_AUTH_KEY")
-
-	switch strings.ToLower(name) {
-	case "otlp", "jaeger":
-		exporter, err = buildOtlpExporter(a.Logger(), name, url, host, port, authHeader)
-	case "zipkin":
-		exporter, err = buildZipkinExporter(a.Logger(), url, host, port, authHeader)
-	case gofrTraceExporter:
-		exporter = buildGoFrExporter(a.Logger(), url)
-	default:
-		a.container.Errorf("unsupported TRACE_EXPORTER: %s", name)
-	}
-
-	return exporter, err
-}
-
-// buildOpenTelemetryProtocol using OpenTelemetryProtocol as the trace exporter
-// jaeger accept OpenTelemetry Protocol (OTLP) over gRPC to upload trace data.
-func buildOtlpExporter(logger logging.Logger, name, url, host, port, authHeader string) (sdktrace.SpanExporter, error) {
-	if url == "" {
-		url = fmt.Sprintf("%s:%s", host, port)
-	}
-
-	logger.Infof("Exporting traces to %s at %s", strings.ToLower(name), url)
-
-	opts := []otlptracegrpc.Option{otlptracegrpc.WithInsecure(), otlptracegrpc.WithEndpoint(url)}
-
-	if authHeader != "" {
-		opts = append(opts, otlptracegrpc.WithHeaders(map[string]string{"Authorization": authHeader}))
-	}
-
-	return otlptracegrpc.New(context.Background(), opts...)
-}
-
-func buildZipkinExporter(logger logging.Logger, url, host, port, authHeader string) (sdktrace.SpanExporter, error) {
-	if url == "" {
-		url = fmt.Sprintf("http://%s:%s/api/v2/spans", host, port)
-	}
-
-	logger.Infof("Exporting traces to zipkin at %s", url)
-
-	var opts []zipkin.Option
-	if authHeader != "" {
-		opts = append(opts, zipkin.WithHeaders(map[string]string{"Authorization": authHeader}))
-	}
-
-	return zipkin.New(url, opts...)
-}
-
-func buildGoFrExporter(logger logging.Logger, url string) sdktrace.SpanExporter {
-	if url == "" {
-		url = "https://tracer-api.gofr.dev/api/spans"
-	}
-
-	logger.Infof("Exporting traces to GoFr at %s", gofrTracerURL)
-
-	return NewExporter(url, logging.NewLogger(logging.INFO))
-}
-
-type otelErrorHandler struct {
-	logger logging.Logger
-}
-
-func (o *otelErrorHandler) Handle(e error) {
-	o.logger.Error(e.Error())
 }
 
 // Subscribe registers a handler for the given topic.
